move FK code out of ObjectBuilder
