+++ tests/Propel/Tests/Generator/Builder/Om/ObjectBuilder/RelationCodeProducer/TernaryNamedRelationCodeTest.php
@@ -8,18 +8,18 @@
 
 namespace Propel\Tests\Generator\Builder\Om\RelationCodeProducer;
 
-use Propel\Generator\Builder\Om\ObjectBuilder\RelationCodeProducer\CrossRelationPartialCodeProducer;
+use Propel\Generator\Builder\Om\ObjectBuilder\RelationCodeProducer\TernaryRelationCodeProducer;
 
 /**
  */
-class CrossRelationBuilderTernaryNamedTest extends AbstractCrossRelationBuilderTest
+class TernaryNamedRelationCodeTest extends AbstractManyToManyCodeTest
 {
     /**
      * @return void
      */
     public function testType(): void
     {
-        $this->assertInstanceOf(CrossRelationPartialCodeProducer::class, $this->getCodeProducer());
+        $this->assertInstanceOf(TernaryRelationCodeProducer::class, $this->getCodeProducer());
     }
 
     /**
